+++ package.json
@@ -18,7 +18,7 @@
     "@groceristar/f-loopback": "^0.0.2",
     "@groceristar/f-showcase": "^0.0.2",
     "@groceristar/fetch-grocery-list": "^0.1.4",
-    "@groceristar/groceristar-fetch": "^2.1.10",
+    "@groceristar/groceristar-fetch": "^2.0.14",
     "@groceristar/recipes-methods": "^0.1.4",
     "body-parser": "^1.19.0",
     "express": "^4.16.4",
